requirements.txt
django==5.1
pillow==10.4.0
psycopg2==2.9.9
djangorestframework==3.15.2
python-dotenv==1.0.1
django-filter==24.3
sqlparse==0.5.1
djangorestframework-simplejwt==5.3.1
coverage==7.6.1
coreapi==2.3.3
drf-yasg==1.21.7
setuptools==73.0.1
stripe==10.8.0
celery==5.4.0
eventlet==0.36.1
redis==5.0.8
django-celery-beat==2.7.0
django-cors-headers==4.4.0
requests==2.32.3